replace with import/export, not working input
