+++ .vscode/launch.json
@@ -1,0 +1,11 @@
+{
+    "configurations": [
+    {
+        "name": "Launch Edge",
+        "request": "launch",
+        "type": "msedge",
+        "url": "http://localhost:5500",
+        "webRoot": "${workspaceFolder}"
+    }
+    ]
+}
