+++ index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8">
   <title>Document</title>
   <link rel="stylesheet" href="style.css">
-  <script type="text/javascript" src="app.js"></script>
+  <script type="module" src="./app.js"></script>
 </head>
 
 <body>
